requirements.txt
youtube-transcript-api>=0.6.0
openai>=1.12.0
python-dotenv>=1.0.1
flask>=3.0.0

